requirements.txt
grad_cam==1.4.8
imutils==0.5.4
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.0.1.24
pandas==2.1.0
Pillow==10.0.0
scipy==1.5.4
seaborn==0.12.2
skimage==0.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
